Drop bootstrap script into run.ts
